feat(util): Mark models to be exported or not
When adding models to a project, they can be marked to be exported or
not using the export=True|False argument. Models with export=false are
not exported, for example, for models still under development.

Also, optional arguments must be passed as keyword arguments now, to
make values and their purpose more explicit.

Fix a small issues with ModelFunc in add being optional, which was left
over from when add worked as a decorator function too.